Update README.md
This project consists of a CRUD (Create, Read, Update, Delete) system for user registration developed with the React library. In the first part is the functionality to add a new user, where the name and username are typed. The second part is the list of registered users with an input field for updating the username and removing the user from the system.
The main focus of this project is the use of the Redux library to manage the state control of the application.